Project/backend/requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
filelock==3.25.2
flatbuffers==25.12.19
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
numpy==2.4.4
onnxruntime==1.24.4
packaging==26.0
protobuf==7.34.1
Pygments==2.20.0
PyYAML==6.0.3
rich==14.3.4
shellingham==1.5.4
sympy==1.14.0
tqdm==4.67.3
typer==0.24.1
typing_extensions==4.15.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.9

langchain-google-genai
langchain-core
python-dotenv
